+++ CHANGES.rst
@@ -5,8 +5,17 @@ Changelog
 1.12 (unreleased)
 -----------------
 
-- Nothing changed yet.
-
+- Added `ram.cache` for `CollectionCategoryVocabulary` to avoid query at each
+  request as query is using a `sort_on=getObjPositionInParent` which is very slow.
+  [gbastien]
+- Override `DashboardCollection` query computation that by default in
+  `plone.app.contenttypes` and `plone.app.querystring` will arbitrary add a `path`
+  index in the query, and again, the `path` index is very slow and just not necessary.
+  [gbastien]
+- Use unrestricted catalog query when possible.
+  [gbastien]
+- Optimized the catalog query that compute the `number_of_items` counters.
+  [gbastien]
 
 1.11 (2020-05-28)
 -----------------